+++ jwt_allauth/registration/views.py
@@ -13,12 +13,15 @@ from rest_framework import status
 from rest_framework.generics import CreateAPIView  #, ListAPIView, GenericAPIView
 # from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
+from django.http import HttpResponseNotFound
 
 # from jwt_allauth.login.views import LoginView
 from jwt_allauth.tokens.models import TokenModel
 from jwt_allauth.registration.app_settings import register_permission_classes
 from jwt_allauth.app_settings import RegisterSerializer
 from jwt_allauth.tokens.app_settings import RefreshToken
+from jwt_allauth.permissions import RegisterUsersPermission
+from jwt_allauth.registration.serializers import UserRegisterSerializer
 # from jwt_allauth.registration.serializers import (
 #     SocialLoginSerializer, SocialAccountSerializer, SocialConnectSerializer)
 from jwt_allauth.utils import get_user_agent, sensitive_post_parameters_m
@@ -52,6 +55,9 @@ class RegisterView(CreateAPIView):
 
     @get_user_agent
     def create(self, request, *args, **kwargs):
+        # If admin-managed registration is enabled, disable open registration endpoint
+        if getattr(settings, 'JWT_ALLAUTH_ADMIN_MANAGED_REGISTRATION', False):
+            return HttpResponseNotFound()
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         token = self.perform_create(serializer)
@@ -71,6 +77,40 @@ class RegisterView(CreateAPIView):
                         allauth_settings.EMAIL_VERIFICATION,
                         None)
         return refresh
+
+
+class UserRegisterView(CreateAPIView):
+    """
+    Admin-managed registration endpoint.
+    - Only accessible to users with admin role (see AdminPermission).
+    - Does not issue tokens on creation.
+    - Triggers email verification; user will set password after verifying.
+    """
+    serializer_class = UserRegisterSerializer
+    permission_classes = (RegisterUsersPermission,)
+    http_method_names = ['post', 'head', 'options']
+
+    @staticmethod
+    def get_response_data(_):
+        return {}
+
+    @sensitive_post_parameters_m
+    def dispatch(self, *args, **kwargs):
+        if not getattr(settings, 'JWT_ALLAUTH_ADMIN_MANAGED_REGISTRATION', False):
+            return HttpResponseNotFound()
+        return super(UserRegisterView, self).dispatch(*args, **kwargs)
+
+    @get_user_agent
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        self.perform_create(serializer)
+        headers = self.get_success_headers(serializer.data)
+        return Response(self.get_response_data(None), status=status.HTTP_201_CREATED, headers=headers)
+
+    def perform_create(self, serializer):
+        serializer.save(self.request)
+        return None
 
 
 # class SocialLoginView(LoginView):